[2~4단계 방탈출 결제 / 배포] 웨이드(고선제) 미션 제출합니다.  (#315)
* refactor: 테스트 위치 수정

* feat: 토스 결제 Client 타임아웃 설정

* test: 타임아웃 테스트 작성

* feat: 내 예약 목록 중 결제 정보 추가

* feat: 결제 저장 로직 구현

* feat: 요청 및 응답 로깅 작성

* feat: 예외 처리 로깅 작성

* refactor : reformating

* feat: 예약 생성 API 로깅 작성

* docs: 배포 스크립트 작성

* docs: 배포 스크립트 수정

* test: 테스트 에러수정

* test: 배포 스크립트 수정

* refactor: 배포스크립트 수정

* refactor: 배포스크립트 수정

* refactor: 배포스크립트 수정

* refactor: 배포스크립트 수정

* refactor: 의미없는 Auth 인터셉터 설정 제거

* refactor: 로깅 인터셉터 가장 높은 우선순위로 명시

* refactor: 타임아웃 발생시 로그 수준 수정

* refactor: 최근 결제 정보를 생성된 날짜로 찾도록 수정

+++ src/test/java/roomescape/presentation/api/dto/request/UserCreateReservationRequestTest.java
@@ -50,7 +50,7 @@ class UserCreateReservationRequestTest {
 
         String paymentKey = "paymentKey";
         String orderId = "id";
-        Long amount = 1234L;
+        int amount = 1234;
         UserCreateReservationRequest request = new UserCreateReservationRequest(date, timeId, themeId, paymentKey,
                 orderId, amount);
 
@@ -65,7 +65,7 @@ class UserCreateReservationRequestTest {
         //given
         String paymentKey = "paymentKey";
         String orderId = "id";
-        Long amount = 1234L;
+        int amount = 1234;
         UserCreateReservationRequest request = new UserCreateReservationRequest(
                 LocalDate.now().plusDays(1),
                 1L,
